Extract everything needed to build an update for a single group into a mixin
